+++ CHANGELOG.md
@@ -4,6 +4,14 @@ The format of this file is based on [Keep a Changelog], and this
 project uses [Semantic Versioning].
 
 
+## [Unreleased]
+
+### Fixed
+
+- `geoextract.WindowExtractor` and its subclasses ignored the first word of
+  some strings.
+
+
 ## [0.2.0] (2017-05-31)
 
 ### Added
@@ -45,6 +53,7 @@ First release.
 [Keep a Changelog]: http://keepachangelog.com
 [Semantic Versioning]: http://semver.org/
 
+[Unreleased]: https://github.com/stadt-karlsruhe/geoextract/compare/v0.2.0...master
 [0.2.0]: https://github.com/stadt-karlsruhe/geoextract/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/stadt-karlsruhe/geoextract/commits/v0.1.0